+++ packages/jivs-engine/src/DataTypes/DataTypeCheckGenerators.ts
@@ -1,6 +1,4 @@
 /**
- * Concrete implementations of IDataTypeCheckGenerator.
- * 
  * {@inheritDoc DataTypes/Types/IDataTypeCheckGenerator!IDataTypeCheckGenerator:interface }
  * @module DataTypes/ConcreteClasses/DataTypeCheckGenerators
  */