+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.17.0
+- **Major feature** Added a parser feature to convert from input value to native value on calls to InputValueHost.setInputValue. It includes a service (DataTypeParserService), interface (IDataTypeParser), and classes to cover a few common cases.
+Aside from using it with the client-side, this will help when posting back form data, to convert the strings sent from the client into native values ready to be validated.
 ## 0.16.2
 - Introduce IDisposable interface to allow more direct control over releasing ValidationManager and ValidationServices.
 - Introduce these Conditions all evaluating numbers: PositiveCondition, IntegerCondition, MaxDecimalsCondition